Refactor and organsize
