+++ src/ZenonWalletApi/Infrastructure/Configurations/SerilogConfigurator.cs
@@ -1,0 +1,25 @@
+﻿using Serilog;
+using Serilog.Core;
+
+namespace ZenonWalletApi.Infrastructure.Configurations
+{
+    public static class SerilogConfigurator
+    {
+        public static Logger CreateLogger()
+        {
+            var configuration = LoadAppConfiguration();
+            return new LoggerConfiguration()
+                .ReadFrom.Configuration(configuration)
+                .CreateLogger();
+        }
+
+        private static IConfigurationRoot LoadAppConfiguration()
+        {
+            return new ConfigurationBuilder()
+                .SetBasePath(Directory.GetCurrentDirectory())
+                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+                .AddJsonFile($"appsettings.{Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") ?? "Production"}.json", optional: true)
+                .Build();
+        }
+    }
+}
