+++ src/ZenonWalletApi/Infrastructure/ExceptionHandlers/RemoteInvocationExceptionHandler.cs
@@ -1,36 +1,32 @@
 ﻿using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
+using StreamJsonRpc;
 using System.Net;
-using System.Net.Sockets;
-using System.Net.WebSockets;
-using Zenon.Client;
 
-namespace ZenonWalletApi.Services.ExceptionHandlers
+namespace ZenonWalletApi.Infrastructure.ExceptionHandlers
 {
-    internal class SocketExceptionHandler : IExceptionHandler
+    internal class RemoteInvocationExceptionHandler : IExceptionHandler
     {
-        public SocketExceptionHandler(ILogger<SocketExceptionHandler> logger)
+        public RemoteInvocationExceptionHandler(ILogger<RemoteInvocationExceptionHandler> logger)
         {
             Logger = logger;
         }
 
-        private ILogger<SocketExceptionHandler> Logger { get; }
+        private ILogger<RemoteInvocationExceptionHandler> Logger { get; }
 
         public async ValueTask<bool> TryHandleAsync(
             HttpContext httpContext,
             Exception exception,
             CancellationToken cancellationToken)
         {
-            if (exception is not SocketException ||
-                exception is not WebSocketException ||
-                exception is not NoConnectionException)
+            if (exception is not RemoteInvocationException)
             {
                 return false;
             }
 
-            Logger.LogError(exception, "Connection exception");
+            Logger.LogError(exception, "Remote invocation exception");
 
-            httpContext.Response.StatusCode = (int)HttpStatusCode.NotFound;
+            httpContext.Response.StatusCode = (int)HttpStatusCode.Conflict;
             await httpContext.Response.WriteAsJsonAsync(new ProblemDetails
             {
                 Status = httpContext.Response.StatusCode,
