+++ src/ZenonWalletApi/Options/ApiOptions.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+using ZenonWalletApi.Models;
+
+namespace ZenonWalletApi.Options
+{
+    public class ApiOptions
+    {
+        public const string Api = "Api";
+
+        [Required]
+        public required IEnumerable<User> Users { get; set; }
+    }
+}
