+++ src/ZenonWalletApi/Program.cs
@@ -1,728 +1,29 @@
-using FluentValidation;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
-using System.Numerics;
-using System.Text;
-using Zenon;
-using Zenon.Model.Embedded.Json;
-using Zenon.Model.NoM;
-using Zenon.Model.NoM.Json;
-using Zenon.Model.Primitives;
-using Zenon.Utils;
+using Serilog;
 using ZenonWalletApi;
-using ZenonWalletApi.Authorization;
-using ZenonWalletApi.Models;
-using ZenonWalletApi.Models.Converters;
-using ZenonWalletApi.Models.Exceptions;
-using ZenonWalletApi.Models.Parameters;
-using ZenonWalletApi.Services;
-using ZenonWalletApi.Services.ExceptionHandlers;
+using ZenonWalletApi.Infrastructure.Configurations;
 
-var builder = WebApplication.CreateBuilder(args);
+Log.Logger = SerilogConfigurator.CreateLogger();
 
-builder.Services.AddHttpClient<IPlasmaBotService, PlasmaBotService>();
-
-// Configuration
-builder.Services.AddOptions<ApiOptions>()
-    .BindConfiguration(ApiOptions.Api)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<JwtOptions>()
-    .BindConfiguration(JwtOptions.Jwt)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<NodeOptions>()
-    .BindConfiguration(NodeOptions.Node)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<WalletOptions>()
-    .BindConfiguration(WalletOptions.Wallet)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<AutoReceiveOptions>()
-    .BindConfiguration(AutoReceiveOptions.AutoReceiver)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<AutoLockServiceOptions>()
-    .BindConfiguration(AutoLockServiceOptions.AutoLocker)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-builder.Services.AddOptions<PlasmaBotOptions>()
-    .BindConfiguration(PlasmaBotOptions.PlasmaBot)
-    .ValidateDataAnnotations()
-    .ValidateOnStart();
-
-// Services
-builder.Services.AddScoped<IJwtService, JwtService>();
-builder.Services.AddScoped<IUserService, UserService>();
-builder.Services.AddTransient<INodeService, NodeService>();
-builder.Services.AddSingleton<IWalletService, WalletService>();
-builder.Services.AddSingleton<IAutoReceiverService, AutoReceiverService>();
-builder.Services.AddSingleton<IAutoLockService, AutoLockService>();
-
-builder.Services.AddHostedService(p => p.GetRequiredService<IWalletService>());
-builder.Services.AddHostedService(p => p.GetRequiredService<IAutoReceiverService>());
-builder.Services.AddHostedService(p => p.GetRequiredService<IAutoLockService>());
-
-builder.Services.AddSingleton<IAuthorizationHandler, UserRoleRequirementAuthorizationHandler>();
-
-// Validation
-builder.Services.AddValidatorsFromAssemblyContaining<Program>(ServiceLifetime.Singleton);
-
-builder.Services.ConfigureHttpJsonOptions(options =>
+try
 {
-    options.SerializerOptions.Converters.Add(new AddressJsonConverter());
-    options.SerializerOptions.Converters.Add(new TokenStandardJsonConverter());
-    options.SerializerOptions.Converters.Add(new HashJsonConverter());
-});
-
-builder.Services.AddSwaggerGen(options =>
+    Log.Logger.Information("Starting up");
+    using var webHost = CreateWebHostBuilder(args).Build();
+    await webHost.RunAsync();
+}
+catch (Exception ex)
 {
-    options.MapType<Address>(() => new OpenApiSchema { Type = "string" });
-    options.MapType<TokenStandard>(() => new OpenApiSchema { Type = "string" });
-    options.MapType<Hash>(() => new OpenApiSchema { Type = "string" });
-
-    options.SwaggerDoc("v1", new OpenApiInfo()
-    {
-        Version = "v1",
-        Title = "Zenon Wallet API",
-        Description = "A .NET based cross-platform Wallet API for interacting with Zenon Alphanet - Network of Momentum Phase 1",
-        Contact = new OpenApiContact()
-        {
-            Name = "Zenon Network",
-            Url = new Uri("https://zenon.network")
-        },
-        License = new OpenApiLicense()
-        {
-            Name = "MIT",
-            Url = new Uri("https://opensource.org/licenses/MIT")
-        }
-    });
-    options.AddSecurityDefinition(JwtBearerDefaults.AuthenticationScheme, new OpenApiSecurityScheme()
-    {
-        Name = "Authorization",
-        Type = SecuritySchemeType.ApiKey,
-        Scheme = JwtBearerDefaults.AuthenticationScheme,
-        BearerFormat = "JWT",
-        In = ParameterLocation.Header,
-        Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
-    });
-    options.AddSecurityRequirement(new OpenApiSecurityRequirement()
-    {
-        {
-            new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference
-                {
-                    Type = ReferenceType.SecurityScheme,
-                    Id = JwtBearerDefaults.AuthenticationScheme
-                }
-            },
-            new string[] {}
-        }
-    });
-});
-
-if (!builder.Environment.IsDevelopment())
+    Log.Logger.Fatal(ex, "Application start-up failed");
+    throw;
+}
+finally
 {
-    builder.Services.AddHttpsRedirection(options =>
-    {
-        options.RedirectStatusCode = StatusCodes.Status308PermanentRedirect;
-        options.HttpsPort = 443;
-    });
+    Log.CloseAndFlush();
 }
 
-builder.Services.AddAuthentication(options =>
-{
-    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-    options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-}).AddJwtBearer(options =>
-{
-    var authOptions = builder.Configuration
-        .GetSection(JwtOptions.Jwt)
-        .Get<JwtOptions>()!;
-
-    options.TokenValidationParameters = new TokenValidationParameters
-    {
-        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(authOptions.Secret)),
-        ValidateIssuer = true,
-        ValidateAudience = true,
-        ValidateIssuerSigningKey = true,
-        ValidAudience = authOptions.ValidAudience,
-        ValidIssuer = authOptions.ValidIssuer,
-        ValidateLifetime = authOptions.ExpiresOn.HasValue || authOptions.ExpiresAfter.HasValue,
-        // set clock skew to zero so tokens expire exactly at token expiration time (instead of 5 minutes later)
-        ClockSkew = TimeSpan.Zero
-    };
-});
-builder.Services.AddAuthorization(options =>
-{
-    options.AddPolicy("Admin", policy =>
-    {
-        policy.AddAuthenticationSchemes(JwtBearerDefaults.AuthenticationScheme);
-        policy.RequireAuthenticatedUser();
-        policy.Requirements.Add(new UserRoleRequirement(UserRoles.Admin));
-    });
-    options.AddPolicy("User", policy =>
-    {
-        policy.AddAuthenticationSchemes(JwtBearerDefaults.AuthenticationScheme);
-        policy.RequireAuthenticatedUser();
-        policy.Requirements.Add(new UserRoleRequirement(UserRoles.User));
-    });
-});
-builder.Services.AddEndpointsApiExplorer();
-
-// Exception handlers
-builder.Services.AddExceptionHandler<IncorrectPasswordExceptionHandler>();
-builder.Services.AddExceptionHandler<NotFoundExceptionHandler>();
-builder.Services.AddExceptionHandler<RemoteInvocationExceptionHandler>();
-builder.Services.AddExceptionHandler<SocketExceptionHandler>();
-builder.Services.AddExceptionHandler<WalletExceptionHandler>();
-builder.Services.AddProblemDetails();
-
-var app = builder.Build();
-
-app.UseHttpsRedirection();
-app.UseAuthentication();
-app.UseAuthorization();
-
-app.UseStatusCodePages(async statusCodeContext
-    => await Results.Problem(statusCode: statusCodeContext.HttpContext.Response.StatusCode)
-                 .ExecuteAsync(statusCodeContext.HttpContext));
-
-app.UseExceptionHandler();
-
-app.UseSwagger();
-app.UseSwaggerUI(c =>
-{
-    c.SwaggerEndpoint("/swagger/v1/swagger.json", "Zenon Wallet API V1");
-});
-
-var root = app.MapGroup("/api");
-root.AddEndpointFilterFactory(ValidationFilter.ValidationFilterFactory);
-root.RequireAuthorization();
-
-#region Users Endpoints
-
-var usersGroup = root.MapGroup("/users")
-    .WithTags("Users");
-
-usersGroup.MapPost("/authenticate", async (
-    [FromServices] IUserService users,
-    [Validate] AuthenticateRequest request) =>
-{
-    var result = await users.AuthenticateAsync(request);
-
-    if (result == null)
-    {
-        return Results.Unauthorized();
-    }
-
-    return Results.Ok(result!);
-})
-    .WithName("AuthenticateUser")
-    .WithDescription("Authenticates an user")
-    .Produces<AuthenticateResponse>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .ProducesValidationProblem()
-    .AllowAnonymous();
-
-#endregion
-
-#region Wallet Endpoints
-
-var walletGroup = root.MapGroup("/wallet")
-    .WithTags("Wallet");
-
-walletGroup.MapGet("/status", (
-    [FromServices] IWalletService service) =>
-{
-    return new WalletStatusResponse(service.IsInitialized, service.IsUnlocked);
-})
-    .WithName("GetWalletStatus")
-    .WithDescription("Gets the wallet status indicating whether the wallet is initialized and unlocked")
-    .Produces<WalletStatusResponse>()
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .RequireAuthorization("User");
-
-walletGroup.MapPost("/{accountIndex}/address", async (
-    [FromServices] IWalletService wallet,
-    [Validate] AccountIndex accountIndex) =>
-{
-    var address = await wallet.GetAccountAddressAsync(accountIndex.value);
-
-    return new WalletAccountAddressResponse(address);
-})
-    .WithName("GetWalletAccountAddress")
-    .WithDescription("Gets the wallet account address by account index")
-    .Produces<WalletAccountAddressResponse>()
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem()
-    .RequireAuthorization("User");
-
-walletGroup.MapPost("/init", async (
-    [FromServices] IWalletService service,
-    [Validate] InitWalletRequest request) =>
-{
-    var mnemonic = await service.InitAsync(request.Password);
-
-    return new InitWalletResponse(mnemonic);
-})
-    .WithName("InitializeWallet")
-    .WithDescription("Initializes a new wallet")
-    .Produces<InitWalletResponse>()
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem()
-    .RequireAuthorization("Admin");
-
-walletGroup.MapPost("/restore", async (
-    [FromServices] IWalletService service,
-    [Validate] RestoreWalletRequest request) =>
-{
-    await service.RestoreAsync(request.Password, request.Mnemonic);
-
-    return Results.Ok();
-})
-    .WithName("RestoreWallet")
-    .WithDescription("Restores an existing wallet")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem()
-    .RequireAuthorization("Admin");
-
-walletGroup.MapPost("/unlock", async (
-    [FromServices] IWalletService service,
-    [Validate] UnlockWalletRequest request) =>
-{
-    await service.UnlockAsync(request.Password);
-
-    return Results.Ok();
-})
-    .WithName("UnlockWallet")
-    .WithDescription("Unlocks the wallet")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem()
-    .RequireAuthorization("User");
-
-walletGroup.MapPost("/lock", async (
-    [FromServices] IWalletService service) =>
-{
-    await service.LockAsync();
-
-    return Results.Ok();
-})
-    .WithName("LockWallet")
-    .WithDescription("Locks the wallet")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .RequireAuthorization("User");
-
-#endregion
-
-#region Ledger Endpoints
-
-var ledgerGroup = root.MapGroup("/ledger")
-    .WithTags("Ledger")
-    .RequireAuthorization("User");
-
-ledgerGroup.MapGet("/{address}/balances", async (
-    [FromServices] INodeService client,
-    [Validate] AddressString address) =>
-{
-    await client.ConnectAsync();
-
-    var result = await client.Api.Ledger.GetAccountInfoByAddress(address.value);
-
-    return result.ToJson();
-})
-    .WithName("GetAccountInfo")
-    .WithDescription("Gets the account height and token balances by address")
-    .Produces<JAccountInfo>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesValidationProblem();
-
-ledgerGroup.MapGet("/{address}/received", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AddressString address,
-    [AsParameters][Validate] TransferReceivedRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Retrieve all received account blocks by address
-    var result = await client.Api.Ledger
-        .GetAccountBlocksByPage(address.value,
-            pageIndex: (uint)request.pageIndex,
-            pageSize: (uint)request.pageSize);
-
-    return result.ToJson();
-})
-    .WithName("GetReceivedAccountBlocks")
-    .WithDescription("Gets the received account blocks by address")
-    .Produces<JAccountBlockList>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesValidationProblem();
-
-ledgerGroup.MapGet("/{address}/unreceived", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AddressString address,
-    [AsParameters][Validate] TransferUnreceivedRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Retrieve all unreceived account blocks by address
-    var result = await client.Api.Ledger
-        .GetUnreceivedBlocksByAddress(address.value,
-            pageIndex: (uint)request.pageIndex,
-            pageSize: (uint)request.pageSize);
-
-    return result.ToJson();
-})
-    .WithName("GetUnreceivedAccountBlocks")
-    .WithDescription("Gets the unreceived account blocks by address")
-    .Produces<JAccountBlockList>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesValidationProblem();
-
-ledgerGroup.MapGet("{address}/plasma", async (
-    [FromServices] INodeService client,
-    [Validate] AddressString address) =>
-{
-    await client.ConnectAsync();
-
-    // Retrieve plasma info
-    var result = await client.Api.Embedded.Plasma.Get(address.value);
-
-    return result.ToJson();
-})
-    .WithName("GetPlasmaInfo")
-    .WithDescription("Gets the current plasma, max plasma and fused qsr amount of the supplied address")
-    .Produces<JPlasmaInfo>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesValidationProblem();
-
-ledgerGroup.MapGet("/{address}/fused", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AddressString address,
-    [AsParameters][Validate] PlasmaFusedRequest request) =>
-{
-    await client.ConnectAsync();
-
-    var syncInfo = await client.Api.Stats.SyncInfo();
-
-    // Retrieve plasma info
-    var result = await client.Api.Embedded.Plasma
-        .GetEntriesByAddress(address.value, (uint)request.pageIndex, (uint)request.pageSize);
-
-    var json = result.ToJson();
-
-    foreach (var element in json.list)
-    {
-        element.isRevocable = syncInfo.currentHeight > element.expirationHeight;
-    }
-
-    return json;
-})
-    .WithName("GetFusionInfo")
-    .WithDescription("Gets the fusion entries of the supplied address")
-    .Produces<JFusionEntryList>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesValidationProblem();
-
-#endregion
-
-#region Transfer Endpoints
-
-var transferGroup = root.MapGroup("/transfer")
-    .WithTags("Transfer")
-    .RequireAuthorization("User");
-
-transferGroup.MapPost("/{accountIndex}/send", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AccountIndex accountIndex,
-    [FromBody][Validate] SendTransferRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Access wallet account
-    var account = await wallet.GetAccountAsync(accountIndex.value);
-
-    // Get wallet account address
-    var address = await account.GetAddressAsync();
-
-    BigInteger amount;
-    if (request.TokenStandard == TokenStandard.ZnnZts ||
-        request.TokenStandard == TokenStandard.QsrZts)
-    {
-        amount = AmountUtils.ExtractDecimals(request.Amount, Constants.CoinDecimals);
-    }
-    else
-    {
-        var token = await client.Api.Embedded.Token.GetByZts(request.TokenStandard);
-        if (token == null)
+static IHostBuilder CreateWebHostBuilder(string[] args) =>
+    Host.CreateDefaultBuilder(args)
+        .UseSerilog()
+        .ConfigureWebHostDefaults(webBuilder =>
         {
-            throw new NotFoundException("Token does not exist");
-        }
-        amount = AmountUtils.ExtractDecimals(request.Amount, (int)token.Decimals);
-    }
-
-    // Retrieve account info
-    var accountInfo = await client.Api.Ledger
-        .GetAccountInfoByAddress(address);
-
-    // Find balance info
-    var balanceInfo = accountInfo.BalanceInfoList
-        .FirstOrDefault(x => x.Token.TokenStandard == request.TokenStandard);
-
-    // Check balance
-    if (balanceInfo == null)
-    {
-        throw new NotFoundException($"You do not have any {request.TokenStandard} tokens");
-    }
-    else if (balanceInfo.Balance < amount)
-    {
-        if (balanceInfo.Balance == BigInteger.Zero)
-        {
-            throw new NotFoundException($"You do not have any {balanceInfo.Token.Symbol} tokens");
-        }
-        else
-        {
-            throw new NotFoundException($"You do not have enough {balanceInfo.Token.Symbol} tokens");
-        }
-    }
-
-    // Create send block
-    var block = AccountBlockTemplate.Send(
-        client.ProtocolVersion, client.ChainIdentifier,
-        request.Address, request.TokenStandard, amount);
-
-    // Send block
-    var response = await client.SendAsync(block, account);
-
-    // Return block hash
-    return response.ToJson();
-})
-    .WithName("SendTransfer")
-    .WithDescription("Sends tokens to an address")
-    .Produces<JAccountBlockTemplate>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-transferGroup.MapPost("/{accountIndex}/receive", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AccountIndex accountIndex,
-    [FromBody][Validate] ReceiveTransferRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Access wallet account from index
-    var account = await wallet.GetAccountAsync(accountIndex.value);
-
-    // Create receive block
-    var block = AccountBlockTemplate.Receive(
-        client.ProtocolVersion, client.ChainIdentifier,
-        request.BlockHash);
-
-    // Send block
-    var response = await client.SendAsync(block, account);
-
-    return response.ToJson();
-})
-    .WithName("ReceiveTransfer")
-    .WithDescription("Manually receives an account block by block hash")
-    .Produces<JAccountBlockTemplate>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-#endregion
-
-#region Plasma Endpoints
-
-var plasmaGroup = root.MapGroup("/plasma")
-    .WithTags("Plasma")
-    .RequireAuthorization("User");
-
-plasmaGroup.MapPost("/{accountIndex}/fuse", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AccountIndex accountIndex,
-    [FromBody][Validate] FusePlasmaRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Access wallet account from index
-    var account = await wallet.GetAccountAsync(accountIndex.value);
-
-    // Retrieve wallet account address
-    var address = await account.GetAddressAsync();
-
-    var amount = AmountUtils.ExtractDecimals(request.Amount, Constants.CoinDecimals);
-
-    // Create block
-    var block = client.Api.Embedded.Plasma.Fuse(request.Address, amount);
-
-    // Send block
-    var response = await client.SendAsync(block, account);
-
-    return response.ToJson();
-})
-    .WithName("FusePlasma")
-    .WithDescription("Fuses QSR to an address to generate plasma")
-    .Produces<JAccountBlockTemplate>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-plasmaGroup.MapPost("/{accountIndex}/cancel", async (
-    [FromServices] IWalletService wallet,
-    [FromServices] INodeService client,
-    [Validate] AccountIndex accountIndex,
-    [FromBody][Validate] CancelPlasmaRequest request) =>
-{
-    await client.ConnectAsync();
-
-    // Access wallet account from index
-    var account = await wallet.GetAccountAsync(accountIndex.value);
-
-    // Retrieve wallet account address
-    var address = await account.GetAddressAsync();
-
-    // Create block
-    var block = client.Api.Embedded.Plasma.Cancel(request.IdHash);
-
-    // Send block
-    var response = await client.SendAsync(block, account);
-
-    return response.ToJson();
-})
-    .WithName("CancelPlasma")
-    .WithDescription("Cancels a plasma fusion and receive the QSR back")
-    .Produces<JAccountBlockTemplate>(StatusCodes.Status200OK)
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-#endregion
-
-#region AutoReceiver Endpoints
-
-var autoReceiverGroup = root.MapGroup("/auto-receiver")
-    .WithTags("AutoReceiver")
-    .RequireAuthorization("User");
-
-autoReceiverGroup.MapGet("/status", (
-    [FromServices] IAutoReceiverService service) =>
-{
-    return new AutoReceiverStatusResponse(service.IsEnabled);
-})
-    .WithName("GetAutoReceiverStatus")
-    .WithDescription("Gets the auto-receiver status indicating whether the service is enabled")
-    .Produces<AutoReceiverStatusResponse>()
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .RequireAuthorization("User");
-
-autoReceiverGroup.MapPut("/{accountIndex}", async (
-    [FromServices] IAutoReceiverService autoReceiver,
-    [Validate] AccountIndex accountIndex) =>
-{
-    await autoReceiver.SubscribeAsync(accountIndex.value);
-
-    return Results.Ok();
-})
-    .WithName("SubscribeAccount")
-    .WithDescription("Subscribes an account to auto receive account blocks")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status404NotFound)
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-autoReceiverGroup.MapDelete("/{accountIndex}", async (
-    [FromServices] IAutoReceiverService autoReceiver,
-    [Validate] AccountIndex accountIndex) =>
-{
-    await autoReceiver.UnsubscribeAsync(accountIndex.value);
-
-    return Results.Ok();
-})
-    .WithName("UnsubscribeAccount")
-    .WithDescription("Unsubscribes an account from auto receiving account blocks")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status409Conflict)
-    .ProducesValidationProblem();
-
-#endregion
-
-#region Utilities EndPoints
-
-var utilsGroup = root.MapGroup("/utilities")
-    .WithTags("Utilities")
-    .RequireAuthorization("User");
-
-utilsGroup.MapPost("/plasma-bot/fuse", async (
-    [FromServices] IPlasmaBotService plasmaBot,
-    [FromBody][Validate] FuseBotPlasmaRequest request) =>
-{
-    return await plasmaBot.FuseAsync(request.Address);
-})
-    .WithName("FuseBotPlasma")
-    .WithDescription("Fuses QSR to an address to generate plasma using the community plasma-bot")
-    .Produces(StatusCodes.Status200OK, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status401Unauthorized, typeof(string), contentType: "text/plain")
-    .Produces(StatusCodes.Status403Forbidden, typeof(string), contentType: "text/plain")
-    .ProducesProblem(StatusCodes.Status400BadRequest)
-    .ProducesProblem(StatusCodes.Status422UnprocessableEntity)
-    .ProducesValidationProblem();
-
-#endregion
-
-app.Run();
+            webBuilder.UseStartup<Startup>();
+        });
