+++ src/ZenonWalletApi/appsettings.Development.json
@@ -1,10 +1,4 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Debug",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
   "Api": {
     "Node": {
       "NodeUrl": "wss://syrius-testnet.zenon.community:443",
